12.2

+++ 12/Main.java
@@ -6,10 +6,10 @@ public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
         System.out.println("enter size");
         int k = sc.nextInt();
-double[] f = new double[k];
+int[] f = new int[k];
 try { 
 for (int x = 0; x < k + 1; x++) {
-    f[x] = (double) x/(x-7);
+    f[x] = (int) x/(x-7);
 System.out.println("x="+x+", f[x]=" + f[x]);
 }
 }
